requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.6.3
async-timeout==4.0.2
attrs==22.2.0
cachetools==4.2.2
certifi==2022.12.7
charset-normalizer==3.1.0
dateparser==1.1.8
frozenlist==1.3.3
idna==3.4
kucoin-python==1.0.11
multidict==6.0.4
numpy==1.24.2
pandas==2.0.0
peewee==3.16.0
psycopg2==2.9.6
pycryptodome==3.17
python-binance==1.0.17
python-dateutil==2.8.2
python-telegram-bot==13.15
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
regex==2023.3.23
requests==2.28.2
six==1.16.0
tornado==6.1
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
urllib3==1.26.15
websockets==11.0
yarl==1.8.2
python-dotenv>=1.0.0
ta==0.10.2